Diplom_2. Исправление замечаний: перенос генерации данных пользователя в helpers.py
