+++ data.py
@@ -24,21 +24,6 @@ class ErrorMessages:
     UNAUTHORIZED = 'You should be authorised'
 
 
-# Данные пользователя
-class UserData:
-
-    # Генерация валидных данных для регистрации пользователя
-    @staticmethod
-    def generation_valid_data_for_create_user():
-        faker = Faker('ru_RU')
-        data = {
-            "email": faker.email(),
-            "password": faker.password(),
-            "name": faker.first_name()
-        }
-        return data
-
-
 # Данные заказа
 class OrderData:
     # Корректные ингредиенты
